Add multiple Parser classes
